feat(j-s): Add prefill text to rulings
Add prefill text to case facts and legal arguments
