+++ apps/judicial-system/web/messages/index.ts
@@ -6,7 +6,6 @@ export * from './RestrictionCases/Prosecutor/caseFiles'
 export * from './RestrictionCases/Prosecutor/overview'
 export * from './/RestrictionCases/Prosecutor/demandsForm'
 export * from './RestrictionCases/Court/courtRecordForm'
-export * from './RestrictionCases/Court/ruling'
 export * from './RestrictionCases/Court/confirmation'
 export * from './RestrictionCases/Court/overview'
 export * from './InvestigationCases/Prosecutor/defendantForm'
@@ -16,7 +15,6 @@ export * from './InvestigationCases/Prosecutor/reportForm'
 export * from './InvestigationCases/Prosecutor/caseFiles'
 export * from './InvestigationCases/Prosecutor/overview'
 export * from './InvestigationCases/Court/courtRecordForm'
-export * from './InvestigationCases/Court/ruling'
 export * from './InvestigationCases/Court/confirmation'
 export * from './InvestigationCases/Court/overview'
 export * from './Indictments/Prosecutor/processing'
